+++ src/com/donny/dendrofinance/data/backingtable/TaxItemBTC.java
@@ -85,6 +85,7 @@ public class TaxItemBTC extends BackingTableCore<TaxItem> {
     @Override
     public void sort() {
         TABLE.sort(Comparator.comparing(taxItem -> taxItem.NAME));
+        changed = true;
     }
 
     @Override